requirements.txt
appium-python-client>=4.0.0
pytest>=8.0.0
python-dotenv>=1.0.1
pytest-html>=4.1.1
